+++ src/main/java/Hogwarts/service/ScenicSpotService.java
@@ -3,6 +3,6 @@ package Hogwarts.service;
 import Hogwarts.entity.ScenicSpot;
 
 public interface ScenicSpotService {
-    ScenicSpot create(String name, String location, float lx, float ly, String description, int hot);
+    ScenicSpot create(String name, String location, double lx, double ly, String description, int hot);
     //void updatehot(int sid);
 }
